CP-54351 move srmetadata from /opt
